begin ReSTful JSON handler for patches, located under /patches/
Returns the list of patches when called without an argument.
Returns the full patch including the documentation and hex data
when called for a single patch, i.e. /patches/1041

Existing blobs are automatically unpacked into slots for patches
that have not yet been converted.  Later on, the blob should be
updated from the slots when the slots have been changed.